requirements.txt
Cython==0.29.21
librosa==0.8.0
matplotlib==3.3.1
numpy==1.18.5
phonemizer==2.2.1
scipy==1.5.2
tensorboard==2.3.0
torchvision
Unidecode==1.1.1
pypinyin
pypinyin_dict
jieba
discord.py
g4f